+++ src/main/java/mcjty/rftools/blocks/environmental/modules/PeacefulEModule.java
@@ -16,7 +16,7 @@ public class PeacefulEModule extends BuffEModule {
 
     @Override
     public float getRfPerTick() {
-        return (float) EnvironmentalConfiguration.REGENERATION_RFPERTICK.get();
+        return (float) (double) EnvironmentalConfiguration.REGENERATION_RFPERTICK.get();
     }
 
     @Override
